RAG_PDF_Reader_Streamlit/requirements.txt
langchain
langchain_openai
python-dotenv
langchain-community
langchain_core
pypdf
langchain-chroma
